+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/repository/IChoreRepository.kt
@@ -6,5 +6,6 @@ import com.altmann.choresmanager.network.ApiResult
 
 interface IChoreRepository {
     suspend fun getChores(userId : Int) : ApiResult<List<ChoreResponseBase>>
+    suspend fun updateChore(userId: Int, chore: ChoreResponseBase) : ApiResult<ChoreResponseBase>
     suspend fun addChore(userId: Int, chore: Chore) : ApiResult<ChoreResponseBase>
 }